dump version

+++ app/Http/Controllers/Api/Client/AuthController.php
@@ -39,7 +39,7 @@ class AuthController extends Controller
         $data = $validator->validated();
 
         // 创建新用户
-        if ($user = Helpers::addUser($data['username'], $data['password'], (int) sysConfig('default_traffic'), sysConfig('default_days'), null, $data['nickname'])) {
+        if ($user = Helpers::addUser($data['username'], $data['password'], (int) sysConfig('default_traffic'), (int) sysConfig('default_days'), null, $data['nickname'])) {
             auth()->login($user, true);
 
             return $this->succeed([
